[FIX] clean up alignment tests
