+++ test/unit/alignment/pairwise/CMakeLists.txt
@@ -1,9 +1,12 @@
 seqan3_test(align_pairwise_test.cpp)
-seqan3_test(alignment_result_test.cpp)
 seqan3_test(align_result_selector_test.cpp)
 seqan3_test(alignment_configurator_test.cpp)
-seqan3_test(edit_distance_unbanded_test.cpp)
+seqan3_test(alignment_result_test.cpp)
 seqan3_test(global_affine_banded_test.cpp)
 seqan3_test(global_affine_unbanded_test.cpp)
+seqan3_test(global_edit_distance_unbanded_test.cpp)
+seqan3_test(semi_global_affine_banded_test.cpp)
+seqan3_test(semi_global_affine_unbanded_test.cpp)
+seqan3_test(semi_global_edit_distance_unbanded_test.cpp)
 
 add_subdirectories()
