+++ test/unit/alignment/pairwise/global_affine_banded_test.cpp
@@ -7,116 +7,58 @@
 
 #include <gtest/gtest.h>
 
+#include <vector>
 
-#include <seqan3/alignment/matrix/alignment_score_matrix.hpp>
-#include <seqan3/alignment/matrix/alignment_trace_matrix.hpp>
-#include <seqan3/alignment/pairwise/align_pairwise.hpp>
+#include <seqan3/alignment/configuration/align_config_band.hpp>
+#include <seqan3/alignment/configuration/align_config_mode.hpp>
+#include <seqan3/alignment/configuration/align_config_gap.hpp>
+#include <seqan3/alignment/configuration/align_config_scoring.hpp>
+#include <seqan3/alignment/scoring/nucleotide_scoring_scheme.hpp>
+#include <seqan3/alignment/scoring/aminoacid_scoring_scheme.hpp>
+#include <seqan3/alphabet/aminoacid/aa27.hpp>
+#include <seqan3/alphabet/nucleotide/dna4.hpp>
+#include <seqan3/alphabet/nucleotide/dna5.hpp>
 
-#include <seqan3/range/view/to_char.hpp>
-
-#include "fixture/global_affine_banded.hpp"
-#include "fixture/semi_global_affine_banded.hpp"
+#include "affine_test_template.hpp"
 
 using namespace seqan3;
 using namespace seqan3::detail;
-using namespace seqan3::test::alignment::fixture;
 
-template <auto _fixture>
-struct param : public ::testing::Test
+namespace seqan3::test::alignment::fixture::global::affine::banded
 {
-    auto fixture() -> decltype(alignment_fixture{*_fixture}) const &
+
+inline constexpr auto align_config = align_cfg::mode{global_alignment} |
+                                     align_cfg::gap{gap_scheme{gap_score{-1}, gap_open_score{-10}}} |
+                                     align_cfg::band{static_band{lower_bound{-3}, upper_bound{8}}};
+
+static auto dna4_01 = []()
+{   //    AACCGGTTAACCGGTT
+    //   01234567890123456|
+    //  0        x        |
+    // A1         x       |
+    // C2          x      |
+    // G3x          x     |
+    // T4 x          x    |
+    // A5  x          x   |
+    // C6   x          x  |
+    // G7    x          x |
+    // T8     x          x|
+    // A9      x          |
+    return alignment_fixture
     {
-        return *_fixture;
-    }
-};
+        "AACCGGTTAACCGGTT"_dna4,
+        "ACGTACGTA"_dna4,
+        align_config | align_cfg::scoring{nucleotide_scoring_scheme{match_score{4}, mismatch_score{-5}}},
+        -18,
+        "A---ACCGGTTAACCGGTT",
+        "ACGTAC----------GTA",
+        alignment_coordinate{detail::column_index_type{0u}, detail::row_index_type{0u}},
+        alignment_coordinate{detail::column_index_type{16u}, detail::row_index_type{9u}}
+    };
+}();
 
-template <typename param_t>
-class global_affine_banded : public param_t
-{};
+} // namespace seqan3::test::alignment::fixture::global::affine::banded
 
-TYPED_TEST_CASE_P(global_affine_banded);
+using global_affine_banded_types = ::testing::Types<param<&global::affine::banded::dna4_01>>;
 
-using global_affine_banded_types
-    = ::testing::Types<
-        param<&global::affine::banded::dna4_01>
-    >;
-
-using semi_global_affine_banded_types
-    = ::testing::Types<
-        param<&semi_global::affine::banded::dna4_semi_seq1_a>,
-        //param<&semi_global::affine::banded::dna4_semi_seq1_b>,
-        param<&semi_global::affine::banded::dna4_semi_seq2_a>,
-        param<&semi_global::affine::banded::dna4_semi_seq2_b>
-    >;
-
-TYPED_TEST_P(global_affine_banded, score)
-{
-    auto const & fixture = this->fixture();
-    // We only compute the score.
-    auto align_cfg = fixture.config | align_cfg::result{with_score};
-
-    std::vector database = fixture.sequence1;
-    std::vector query = fixture.sequence2;
-
-    // TODO Make test work with ranges.
-    auto alignment = align_pairwise(std::tie(database, query), align_cfg);
-
-    EXPECT_EQ((*std::ranges::begin(alignment)).score(), fixture.score);
-}
-
-TYPED_TEST_P(global_affine_banded, end_position)
-{
-    auto const & fixture = this->fixture();
-    auto align_cfg = fixture.config | align_cfg::result{with_back_coordinate};
-
-    std::vector database = fixture.sequence1;
-    std::vector query = fixture.sequence2;
-
-    auto alignment = align_pairwise(std::tie(database, query), align_cfg);
-
-    auto res = *std::ranges::begin(alignment);
-    EXPECT_EQ(res.score(), fixture.score);
-
-    EXPECT_EQ(res.back_coordinate(), fixture.back_coordinate);
-}
-
-TYPED_TEST_P(global_affine_banded, begin_position)
-{
-    auto const & fixture = this->fixture();
-    auto align_cfg = fixture.config | align_cfg::result{with_front_coordinate};
-
-    std::vector database = fixture.sequence1;
-    std::vector query = fixture.sequence2;
-
-    auto alignment = align_pairwise(std::tie(database, query), align_cfg);
-
-    auto res = *std::ranges::begin(alignment);
-    EXPECT_EQ(res.score(), fixture.score);
-
-    EXPECT_EQ(res.front_coordinate(), fixture.front_coordinate);
-    EXPECT_EQ(res.back_coordinate(), fixture.back_coordinate);
-}
-
-TYPED_TEST_P(global_affine_banded, trace)
-{
-    auto const & fixture = this->fixture();
-    auto align_cfg = fixture.config | align_cfg::result{with_alignment};
-
-    std::vector database = fixture.sequence1;
-    std::vector query = fixture.sequence2;
-
-    auto alignment = align_pairwise(std::tie(database, query), align_cfg);
-
-    auto res = *std::ranges::begin(alignment);
-    EXPECT_EQ(res.score(), fixture.score);
-    EXPECT_EQ(res.front_coordinate(), fixture.front_coordinate);
-    EXPECT_EQ(res.back_coordinate(), fixture.back_coordinate);
-    EXPECT_TRUE(ranges::equal(get<0>(res.alignment()) | view::to_char, fixture.aligned_sequence1));
-    EXPECT_TRUE(ranges::equal(get<1>(res.alignment()) | view::to_char, fixture.aligned_sequence2));
-}
-
-REGISTER_TYPED_TEST_CASE_P(global_affine_banded, score, end_position, begin_position, trace);
-
-// work around a bug that you can't specify more than 50 template arguments to ::testing::types
-INSTANTIATE_TYPED_TEST_CASE_P(global, global_affine_banded, global_affine_banded_types);
-INSTANTIATE_TYPED_TEST_CASE_P(semi_global, global_affine_banded, semi_global_affine_banded_types);
+INSTANTIATE_TYPED_TEST_CASE_P(global, align_affine, global_affine_banded_types);
